# field: speed up Element.Bytes