@FiloSottile 2025-02-06
Nice! I appreciate the diagram.

Since these are non-trivial changes to the parts of the library that track upstream, would you consider mailing them as Go CLs? If not, have you signed the Google CLA and are ok with me submitting them on your behalf after we merge them here, to keep upstream in sync?