@AlexanderYastrebov 2025-02-07
> would you consider mailing them as Go CLs?

Sure, I've created https://github.com/golang/go/pull/71603 for this change.

> If not, have you signed the Google CLA and are ok with me submitting them on your behalf after we merge them here, to keep upstream in sync?

Yes, I've contributed to stdlib before so this is also ok.