@AlexanderYastrebov 2025-02-07
https://github.com/golang/go/pull/71603 was landed, how do we proceed with this one?